# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PARDON Anomaly API",
    "description": "Inference, retraining, experiments, and drift notifications API.",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Healthz Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Readyz",
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Readyz Readyz Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Predict",
        "operationId": "predict_api_v1_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/recent": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Recent Predictions",
        "operationId": "recent_predictions_api_v1_predictions_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "model_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecentPrediction"
                  },
                  "title": "Response Recent Predictions Api V1 Predictions Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retrain": {
      "post": {
        "tags": [
          "retrain"
        ],
        "summary": "Retrain",
        "operationId": "retrain_api_v1_retrain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/retrain/{job_id}": {
      "get": {
        "tags": [
          "retrain"
        ],
        "summary": "Retrain Status",
        "operationId": "retrain_status_api_v1_retrain__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Experiments",
        "operationId": "experiments_api_v1_experiments_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "model_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExperimentItem"
                  },
                  "title": "Response Experiments Api V1 Experiments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/drift": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Drift Notifications",
        "operationId": "drift_notifications_api_v1_notifications_drift_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "only_unread",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Only Unread"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriftNotification"
                  },
                  "title": "Response Drift Notifications Api V1 Notifications Drift Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/drift/{notification_id}/read": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Notification Read",
        "operationId": "mark_notification_read_api_v1_notifications_drift__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Mark Notification Read Api V1 Notifications Drift  Notification Id  Read Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DriftNotification": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "severity",
          "title",
          "message",
          "read"
        ],
        "title": "DriftNotification"
      },
      "ExperimentItem": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "experiment": {
            "type": "string",
            "title": "Experiment"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "experiment",
          "model_name",
          "status",
          "started_at",
          "metrics",
          "params"
        ],
        "title": "ExperimentItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PredictRecord": {
        "properties": {
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Features"
          }
        },
        "type": "object",
        "title": "PredictRecord"
      },
      "PredictRequest": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": "isolation_forest"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/PredictRecord"
            },
            "type": "array",
            "minItems": 1,
            "title": "Records"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "records"
        ],
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/PredictionItem"
            },
            "type": "array",
            "title": "Predictions"
          },
          "drift": {
            "additionalProperties": true,
            "type": "object",
            "title": "Drift"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "model_name",
          "predictions",
          "drift"
        ],
        "title": "PredictResponse"
      },
      "PredictionItem": {
        "properties": {
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "anomaly_flag": {
            "type": "integer",
            "title": "Anomaly Flag"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "score",
          "threshold",
          "anomaly_flag"
        ],
        "title": "PredictionItem"
      },
      "RecentPrediction": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "anomaly_flag": {
            "type": "integer",
            "title": "Anomaly Flag"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "request_id",
          "source_id",
          "model_name",
          "score",
          "threshold",
          "anomaly_flag"
        ],
        "title": "RecentPrediction"
      },
      "RetrainJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "dataset_scenario": {
            "type": "string",
            "title": "Dataset Scenario"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "model_name",
          "dataset_scenario",
          "details"
        ],
        "title": "RetrainJobResponse"
      },
      "RetrainRequest": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": "isolation_forest"
          },
          "dataset_scenario": {
            "type": "string",
            "title": "Dataset Scenario",
            "default": "all"
          },
          "register_model": {
            "type": "boolean",
            "title": "Register Model",
            "default": true
          }
        },
        "type": "object",
        "title": "RetrainRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}